+++ src/components/commons/popover/HoverPopover.tsx
@@ -1,0 +1,50 @@
+import { JSX } from '@babel/types'
+import { Transition } from '@headlessui/react'
+import { PropsWithChildren, useState } from 'react'
+import { usePopper } from 'react-popper'
+
+interface IconPopoverProps {
+  description: string
+  className?: string
+}
+
+export function HoverPopover (props: PropsWithChildren<IconPopoverProps>): JSX.Element {
+  const [isHover, setIsHover] = useState(false)
+
+  const [refEle, setRefEle] = useState<any>()
+  const [popperEle, setPopperEle] = useState<any>()
+  const { styles, attributes } = usePopper(refEle, popperEle, { placement: 'bottom' })
+
+  return (
+    <div>
+      <div
+        ref={setRefEle}
+        className={props.className}
+        onMouseOver={() => setIsHover(true)}
+        onMouseLeave={() => setIsHover(false)}
+      >
+        {props.children}
+      </div>
+
+      <Transition
+        enter='transition ease-out duration-200'
+        enterFrom='opacity-0 translate-y-1'
+        enterTo='opacity-100 translate-y-0'
+        leave='transition ease-in duration-150'
+        leaveFrom='opacity-100 translate-y-0'
+        leaveTo='opacity-0 translate-y-1'
+        show
+      >
+        {isHover && (
+          <div ref={setPopperEle} style={styles.popper} {...attributes.popper}>
+            <div className='p-2'>
+              <div className='px-3 py-2 text-sm bg-black text-white rounded shadow-md ring-1 ring-gray-200 max-w-sm'>
+                {props.description}
+              </div>
+            </div>
+          </div>
+        )}
+      </Transition>
+    </div>
+  )
+}
